phase_3/requirements.txt
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
micromlgen==1.1.28
numpy==2.4.2
nvidia-nccl-cu12==2.29.3
pandas==3.0.1
pyserial==3.5
python-dateutil==2.9.0.post0
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
threadpoolctl==3.6.0
xgboost==3.2.0
